Refactor save downloads
